PST parser: can parse coordinates with missing values.

+++ LaTeXDraw/src/net/sf/latexdraw/glib/views/pst/PSTricksConstants.java
@@ -119,7 +119,7 @@ public abstract class PSTricksConstants {
 	public static final double DEFAULT_GRADIENT_ANGLE = 0;
 
 	/** The value by default of a missing coordinate : (5,) */
-	public static final int DEFAULT_VALUE_MISSING_COORDINATE = 1;
+	public static final double DEFAULT_VALUE_MISSING_COORDINATE = 1.;
 
 	public static final String TOKEN_PUT_ANGLE_0   = "U";//$NON-NLS-1$
 	public static final String TOKEN_PUT_ANGLE_90  = "L";//$NON-NLS-1$
